feat(templates): refactor middlewares and services of the Boilerplate project template #6774 (#6775)
